LSP handling of AirCode connection changes.
